\name{DATforDCEMRI-package}
\Rdversion{1.1}
\alias{DATforDCEMRI-package}
\alias{DATforDCEMRI}
\docType{package}
\title{
Deconvolution Analysis Tool for DCE-MRI
}
\description{
Calculates the Impulse Response Function on a per-voxel basis for DCE-MRI data using numerical deconvolution, yielding the Impulse Response Function for each voxel analyzed.  The Area under the Curve and ratio of AUC to Mean Residence Time provide estimates of the kinetic model parameters ve and Ktrans.  A visualization tool is also included that allows the user to explore voxel-wise data in an interactive manner.
}
\details{
\tabular{ll}{
Package: \tab DATforDCEMRI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-03-09\cr
License: \tab What GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Gregory Zelinsky Ferl <ferl.gregory@gene.com>
}
\references{
Ferl GZ, Xu L, Friesenhahn M, Bernstein LJ, Barboriak DP, Port RE. (2010) An automated method for nonparametric kinetic analysis of clinical DCE-MRI data: application to glioblastoma treated with bevacizumab. \emph{Magn Reson Med.} \bold{63}, 1366-1375.
}
\keyword{ package }
